+++ src/app/pages/new-project/new-project.page.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { ModalController } from '@ionic/angular';
+import { Helpers } from 'src/app/helpers/helpers';
 
 @Component({
   selector: 'app-new-project',
@@ -9,15 +11,29 @@ import { Router } from '@angular/router';
 export class NewProjectPage implements OnInit {
 
   projectName: string;
+  existingProjects;
 
-  constructor(private router: Router) { }
+  constructor(private router: Router, private modalCtrl: ModalController, private helpers: Helpers) { }
 
   ngOnInit() {
   }
 
+  onClosePage() {
+    this.modalCtrl.dismiss();
+  }
+
   onNext() {
-    window.localStorage.setItem('projectName', this.projectName);
-    this.router.navigateByUrl('/project');
+    if (this.existingProjects.find(p => p === this.projectName)) {
+      const alert = {
+        head: 'Project Name Already Exists',
+        msg: 'Please enter unique project name'
+      }
+      this.helpers.showAlert(alert.head, alert.msg);
+    } else {
+      window.localStorage.setItem('projectName', this.projectName);
+      this.onClosePage();
+      this.router.navigateByUrl('/project');
+    }
   }
